+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 # Changelog
 
-All Notable changes to `digipolis/toerismevlaanderen-lodging` package.
+All Notable changes to `digipolisgent/toerismevlaanderen-lodging` package.
 
 ## [Unreleased]
 
 Initial package.
 
 ### Added
+
+* Added getting the number of lodges by a given locality.
 
 [1.0.0]: https://github.com/digipolisgent/php_package_dg-toerismevlaanderen-lodging/releases/tag/1.0.0
 [Unreleased]: https://github.com/digipolisgent/php_package_dg-toerismevlaanderen-lodging/compare/master...develop